fix u32 slave_id string parser (#77)
default slave_id  is now unsigned long instead of int.
It means that values greater than 0xF0000000 are now correctly
parsed. Test is added to mercury200 where the problem was first
reported.

(cherry picked from commit 241b83a22a8d6338d8311468ced33a64c52c1275)

+++ src/serial_device.cpp
@@ -189,10 +189,10 @@ PRegisterTypeMap TSerialDeviceFactory::GetRegisterTypes(PDeviceConfig device_con
 }
 
 template<>
-int TBasicProtocolConverter<int>::ConvertSlaveId(const std::string &s) const
+unsigned long TBasicProtocolConverter<unsigned long>::ConvertSlaveId(const std::string &s) const
 {
     try {
-        return std::stoi(s, /* pos = */ 0, /* base = */ 0);
+        return std::stoul(s, /* pos = */ 0, /* base = */ 0);
     } catch (const std::logic_error &e) {
         throw TSerialDeviceException("slave ID \"" + s + "\" is not convertible to string");
     }
